+++ src/WaveRush/Entity/SmartEnemy.cpp
