+++ src/WaveRush/Entity/SmartEnemy.hpp
@@ -1,15 +1,16 @@
 #pragma once
 
-#include "Core/Game.hpp"
-#include "Core/Utils.hpp"
-#include "Entity/Entity.hpp"
+#include "WaveRush/Core/Game.hpp"
+#include "WaveRush/Core/Utils.hpp"
+#include "WaveRush/Entity/Entity.hpp"
 
-class BasicEnemy : public Entity {
+class SmartEnemy : public Entity {
 private:
-	float m_Speed;
+	float m_Attraction;
+	float m_Repulsion;
 
 public:
-	explicit BasicEnemy(
+	explicit SmartEnemy(
 			const glm::vec2& p_Position = glm::vec2(
 					Random<float>(0, Game::Instance().GetSettings().Width),
 					Random<float>(0, Game::Instance().GetSettings().Height)));
